rename bitbox-cli.d.ts to bitbox-sdk.d.ts
